Refactor Node hierarchy modules
