+++ devito/__init__.py
@@ -5,12 +5,11 @@ from sympy.core import cache
 
 from devito.operator import *  # noqa
 from devito.finite_difference import *  # noqa
-from devito.iteration import *  # noqa
 from devito.dimension import *  # noqa
-from devito.expression import *  # noqa
 from devito.stencilkernel import *  # noqa
 from devito.interfaces import *  # noqa
 from devito.interfaces import _SymbolCache
+from devito.nodes import *  # noqa
 
 
 def clear_cache():
