+++ devito/propagator.py
@@ -18,9 +18,8 @@ from devito.compiler import (IntelMICCompiler, get_compiler_from_env,
 from devito.dimension import t, x, y, z
 from devito.dse.inspection import retrieve_dtype
 from devito.dse.symbolics import _temp_prefix, rewrite
-from devito.expression import Expression
 from devito.function_manager import FunctionDescriptor, FunctionManager
-from devito.iteration import Iteration
+from devito.nodes import Expression, Iteration
 from devito.logger import info
 from devito.profiler import Profiler
 from devito.tools import flatten
